refactor: BaseEntity, BaseUpdatableEntity 패키지 이동
